Fix empty string on heading
Sometimes the heading can be an empty string which should be mapped to
None. Note: I really don't like this method of converting erros to None,
it will sliently convert parsing errors to None, which may or may not be
right. I think the best way of handling this would be to log the action
so that we can make sure that the endpoint isn't giving us unexpected
values.